1. 수주 상세페이지 (초기 단계 수정 중)
2. 수주 리스트 수정(요청 수량, 출고 수량, 총액 - 컬럼 추가)
3. 수주 품목의 총 요청 수량, 총 출고 수량, 총 가격 로직 Sales_OrderServiceImpl에 적용
4. Sales_OrderDto 수정( 총 요청 수량, 총 출고 수량, 총 가격 - 필드 추가)

+++ AssemERP/src/main/java/com/WiseForce/AssemERP/dto/km/Sales_ItemDto.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class Sales_ItemDto {
 	private int 		sales_No;
 	private int 		product_No;
-	private Long 		sales_Item_Cnt;
+	private int			sales_Item_Cnt;
 	private int	 		sales_Item_OutCnt;
 	private Long		sales_Item_Cost;
 	private ProductDTO 	productDto;
